// src/app/models/ConsultasItem.ts
export class ConsultasItem {
  idInstitucion: String;
  idConsulta: String;
  idModulo: String;
  general: String;
  nombre: String;
  descripcion: String;
  idObjetivo: String;
  idClaseComunicaciones: String;
  generica: boolean;
  fechaModificacion: Date;
  usuModificacion: String;
  bases: String;
  idTabla: String;
  esExperta: String;
  sentencia: String;
  observaciones: String;
  constructor() { }
}